+++ src/main/java/com/sparta/shop_sparta/config/security/SecurityConfig.java
@@ -1,46 +1,41 @@
 package com.sparta.shop_sparta.config.security;
 
+import com.sparta.shop_sparta.config.security.jwt.JwtAuthenticationFilter;
+import com.sparta.shop_sparta.config.security.jwt.JwtTokenProvider;
 import com.sparta.shop_sparta.util.encoder.SaltGenerator;
 import com.sparta.shop_sparta.util.encoder.UserInformationEncoder;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 @Configuration
 @EnableWebSecurity
+@RequiredArgsConstructor
 public class SecurityConfig {
 
-    @Value("${SECRET_KEY}")
-    public String secretKey;
-
-    @Bean
-    public SaltGenerator saltGenerator() {
-        return new SaltGenerator();
-    }
-
-    // 복호화 가능 암호화 인코더
-    @Bean
-    public UserInformationEncoder userInformationEncoder(){
-        return new UserInformationEncoder(secretKey);
-    }
-
-    // 패스워드 인코더
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder(){
-        return new BCryptPasswordEncoder();
-    }
+    private final JwtTokenProvider jwtTokenProvider;
 
     // 개발 초기 단계 모든 접근 허용
     @Bean
     SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception{
-        http
-                .authorizeHttpRequests( authorize ->
-                        authorize.anyRequest().permitAll()
+        http.
+                authorizeHttpRequests( authorizeRequests ->
+                        authorizeRequests
+                            .anyRequest().permitAll()
                 )
+                .sessionManagement( sessionManagement ->
+                        sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                )
+                .addFilterBefore(new JwtAuthenticationFilter(jwtTokenProvider), UsernamePasswordAuthenticationFilter.class)
+                .formLogin(formLogin -> formLogin.permitAll())
                 .csrf(csrf -> csrf.disable());
 
         return http.build();
